+++ releases.md
@@ -25,6 +25,7 @@ Changes are grouped as follows:
 
 - Utility methods for converting `Value` to various types. This can be useful when working with CDF.Raw which 
 represents its columns as `Struct` and `Value`.
+- Ability to synchronize multiple hierarchies via `Assets.synchronizeMultipleHierarchies(Collection<Asset> assetHierarchies)`.
 - Utility methods for parsing nested `Struct` and `Value` objects. This can be useful when working with CDF.Raw.
 
 ## [0.9.9]